CMAKE - update
* Fix cmake warns - badalloc
* cmake - segfault_sigaction unused var warn
* Fix kat build in cmake
* Remove aalcalc welford method + tidy
* Update build to remove ktgetmem
* Add cmake build docker
* Update gulcalc warns
* fix fmcalc cmake warns

+++ src/fmcalc/fmcalcrules.cpp
@@ -1581,7 +1581,7 @@ void fmcalc::init_profile_rec(fm_profile &f)
 		}
 
 	}
-	if (profile_vec_new_.size() < f.profile_id + 1) {
+	if (profile_vec_new_.size() < (size_t)f.profile_id + 1) {
 		profile_vec_new_.resize(f.profile_id + 1);
 	}
 	profile_vec_new_[f.profile_id] = p;
